requirements.txt
# =============================================================================
# NHS Chest Pain AI Triage System - Python 3.13 Compatible (May 2025)
# Core Medical Triage System WITH Essential Packages
# =============================================================================

# Core Framework & API
fastapi[standard]
uvicorn[standard]
pydantic>=2.0
pydantic-settings
python-multipart
python-jose[cryptography]
passlib[bcrypt]
ollama==0.4.9

# Database & Storage
sqlalchemy>=2.0
alembic
asyncpg
psycopg[binary]
redis
aioredis
chromadb

# Object Storage
minio
boto3

# Machine Learning & AI - Python 3.13 Compatible
transformers>=4.36.0
accelerate
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0

# Computer Vision & Medical Imaging
ultralytics>=8.3.0  # YOLOv8 - latest pre-release supports Python 3.13
opencv-python-headless
Pillow>=10.0.0
imageio

# Data Processing - PANDAS REPLACEMENT
polars>=0.20.0  # Fast DataFrame library, similar pandas syntax
pyarrow>=15.0.0  # Backend for polars

# Medical NLP - Hugging Face Only
nltk
textblob
stanza

# Task Queue & Background Processing
celery[redis]
flower

# Communication APIs
python-telegram-bot>=20.0
twilio

# Document Generation & Reports
reportlab
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
kaleido  # For plotly static exports

# FHIR & Healthcare Standards
fhir.resources
fhirclient

# MCP (Model Context Protocol)
anthropic

# Web Dashboard & UI - MESOP REPLACEMENT
streamlit>=1.32.0  # Much better UI/UX for customer interface
streamlit-plotly-events
streamlit-aggrid  # Advanced data tables
streamlit-authenticator  # User authentication
mesop
# Data Validation & Security
email-validator
python-dateutil
pytz
cryptography
python-dotenv
bcrypt

# Monitoring & Logging
structlog
loguru

# Testing & Development
pytest>=7.4.0
pytest-asyncio
pytest-cov
httpx
faker

# Code Quality
black
ruff
mypy

# Development Tools
ipython
jupyter

# Memory & System Monitoring
psutil
GPUtil

# Healthcare Specific - Python 3.13 Compatible
pyhealth
nltk
textblob

# Infrastructure
docker
redis
